scheduler: revise custom workflow and some helpers
